start Circus if not running, restart if running

+++ src/main/resources/restart.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 p=`pgrep circusd`
-kill -15 ${p}
-while ps -p ${p} 1>/dev/null; do sleep 1; done;
+if [ ${p} ]; then
+  kill -15 ${p}
+  while ps -p ${p} 1>/dev/null; do sleep 1; done;
+fi
 ~/circus/start.sh
